requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.0.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-polymorphic==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.2
notifications==0.3.2
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2023.3
requests==2.30.0
simplejson==3.19.1
sqlparse==0.4.4
swapper==1.3.0
uritemplate==4.1.1
urllib3==2.0.2
